refactor(openai): separate execution operations and transport ownership
