+++ devlog/_plan/260905_production_readiness/044_wp04_build_notes.md
@@ -1,0 +1,41 @@
+# WP04 B — owner ledger
+
+Real B entered after043 independent A synthesis. Original source baseline d039b587;
+docs checkpoint28cdff28. No WP04 PR or completion claim yet.
+
+Main added same-object OpenAI/legacy selection and narrowed legacy provider type,
+preserving the current-key wrapper. Family import will resolve when its assigned
+worker lands. Main also added test-only trackWork and three direct-work settlement
+regressions; these are not yet execution-green while new source imports are absent.
+
+Five bounded workers, each explicitly gpt-6-astra/high, no grandchildren:
+
+| Slice | Agent |
+| --- | --- |
+| Transport/types | Nietzsche01a070a4-b2d9-7312-bf90-c2af975de2c3 |
+| Operations/facade | Anscombe01a070a4-b37a-7ee2-8666-3e4dd20ee230 |
+| Family/legacy extraction | Linnaeus01a070a4-b42b-72e2-ba4c-f3bf29ef9557 |
+| Runtime parity tests | Ramanujan01a070a4-b4e9-7973-828e-d8cd2353afc5 |
+| Contract/mock migration | Cicero01a070a4-b595-7732-bcd8-bf2164d5b88d |
+
+Exact disjoint files and function names are fixed in042. Workers use immutable
+git-show baseline for shared original adapter source, not another worker's edited
+facade. Main coordinates builds/compiler/inventory/source docs and git; no user
+stores, real providers, full local suites or live browser/server are in scope.
+Resource bounds from041 remain in effect; priority is not a separately proven
+spawn field. Independent C and exact-head CI will follow implementation.
+
+## Compiling extraction checkpoint
+
+Main independently compared original d039b587 AST function bodies: normalizeRef
+and all three operations are byte-identical; all ten transport function bodies
+are byte-identical. Transport/types worker also checked exact original type/helper
+blocks and restricted exports. Family worker compared five OpenAI call argument
+lists/classic retry capture and41remaining legacy dependency calls unchanged.
+
+Current server typecheck and build:server exit0. The compatibility file is five
+re-export lines only; current emitted runtime is rebuilt, not committed. Main's
+updated harness13/13 passed, including three new direct-work ownership cases.
+Test compiler initially found a test-factory surface union requiring an explicit
+single-result narrowing; original runtime-test owner is fixing that test-only
+diagnostic. Full new parity and migrated contracts remain in progress.
