+++ lib/providers/execution/legacyNode.ts
@@ -1,4 +1,3 @@
-import { generateViaResponses, editViaResponses } from "../../responsesImageAdapter.js";
 import { generateViaGrok } from "../../grokImageAdapter.js";
 import { generateViaAgy } from "../../agyImageAdapter.js";
 import { generateViaGeminiApi } from "../../geminiApiImageAdapter.js";
@@ -7,31 +6,32 @@ import { generateViaMinimax } from "../../minimaxImageAdapter.js";
 import { generateViaNai } from "../../naiImageAdapter.js";
 import { toGrokReferences } from "../../nodeHelpers.js";
 import { requireRuntimeContext, type RuntimeContext } from "../../runtimeContext.js";
-import type { ExecutionProgress, ImageExecutionRequest, PreparedImageExecution, SingleImageExecutionResult } from "./types.js";
+import type { ExecutionProgress, PreparedImageExecution, SingleImageExecutionResult } from "./types.js";
+import type { LegacyExecutionRequest } from "./legacy.js";
 
-type NodeRequest = Extract<ImageExecutionRequest, { surface: "node" }>;
+type NodeRequest = Extract<LegacyExecutionRequest, { surface: "node" }>;
 
 export async function prepareLegacyNode(
-  ctx: RuntimeContext, request: NodeRequest, progress?: ExecutionProgress,
+  ctx: RuntimeContext, request: NodeRequest, _progress?: ExecutionProgress,
 ): Promise<PreparedImageExecution<"node">> {
   // Preserve the original pre-attempt capture. The facade checks current presence
   // on every execute, but replacing a nonblank key must not rebind either retry.
   const grokDirectApiKey = request.provider === "grok-api" ? ctx.xaiApiKey : undefined;
   return { execute: async () => {
     try {
-      return { kind: "single", value: await executeNodeAttempt(ctx, request, progress, grokDirectApiKey) };
+      return { kind: "single", value: await executeNodeAttempt(ctx, request, grokDirectApiKey) };
     } catch (error) { throw error; } // The caller owns retries and normalization.
   } };
 }
 
 // Keep the legacy one-attempt branches together; retry/lifecycle stays at the caller.
 async function executeNodeAttempt(
-  ctx: RuntimeContext, request: NodeRequest, progress: ExecutionProgress | undefined,
+  ctx: RuntimeContext, request: NodeRequest,
   grokDirectApiKey: string | undefined,
 ): Promise<SingleImageExecutionResult> {
   const { provider, sourceImage: parentB64, prompt: generationPrompt, rawPrompt: prompt,
-    references, requestId, signal, searchMode, options } = request;
-  const { model, size, quality, moderation, mode, reasoningEffort, webSearchEnabled } = options;
+    references, requestId, signal, options } = request;
+  const { model, size, quality } = options;
   const refsForRequest = request.contextMode === "parent-only" ? [] : references;
   return provider === "gemini-api"
     ? await generateViaGeminiApi(parentB64 ? `Edit this image: ${generationPrompt}` : generationPrompt, requireRuntimeContext(ctx), {
@@ -71,17 +71,6 @@ async function executeNodeAttempt(
         references: toGrokReferences(parentB64, refsForRequest),
         directApiKey: grokDirectApiKey,
       })
-    : provider === "api" || provider === "oauth"
-    ? parentB64
-      ? await editViaResponses(provider, generationPrompt, parentB64, quality, size, moderation, mode, ctx, requestId, {
-          model, references: refsForRequest, searchMode, reasoningEffort, webSearchEnabled, signal,
-        })
-      : await generateViaResponses(provider, generationPrompt, quality, size, moderation,
-          refsForRequest, requestId, mode, ctx, {
-            model, reasoningEffort, webSearchEnabled, signal,
-            partialImages: request.partialImages,
-            onPartialImage: progress?.onPartialImage ?? null,
-          })
     : unsupportedNodeProvider(provider);
 }
